Create PBF package and move PBF components from PBEM to PBF
