+++ game-core/src/main/java/org/triplea/game/startup/SetupModel.java
@@ -3,7 +3,7 @@ package org.triplea.game.startup;
 import games.strategy.engine.data.properties.GameProperties;
 import games.strategy.engine.framework.startup.launcher.ILauncher;
 import games.strategy.engine.pbem.IEmailSender;
-import games.strategy.engine.pbem.IForumPoster;
+import games.strategy.engine.pbf.IForumPoster;
 import games.strategy.engine.random.IRemoteDiceServer;
 import java.util.Optional;
 import org.triplea.game.chat.ChatModel;
